+++ db_persistence/migrations/2018-02-18-184901_create_inputs/up.sql
@@ -1,8 +1,8 @@
 CREATE TABLE inputs (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     sequence_number INTEGER NOT NULL,
     previous_tx_hash BYTEA NOT NULL,
     previous_tx_output_index INTEGER NOT NULL,
     script BYTEA NOT NULL,
-    transaction_id INTEGER NOT NULL REFERENCES transactions (id)
+    transaction_id BIGINT NOT NULL REFERENCES transactions (id)
 )
