+++ db_persistence/migrations/2018-02-18-184904_create_outputs/up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE outputs (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     output_index INTEGER NOT NULL,
     value BIGINT NOT NULL,
     script BYTEA NOT NULL,
-    transaction_id INTEGER NOT NULL REFERENCES transactions (id)
+    transaction_id BIGINT NOT NULL REFERENCES transactions (id)
 )
