+++ db_persistence/src/domain/address.rs
@@ -1,7 +1,7 @@
 #[derive(Queryable)]
 pub struct Address {
-  pub id: i32,
+  pub id: i64,
   pub hash: Vec<u8>,
   pub base58check: String,
-  pub output_id: i32,
+  pub output_id: i64,
 }
